AMLS-5522 | Updating specs for new home address split and address package refactor
